requirements.txt
flask
requests
gunicorn
websocket-client
